workspace/paper-repos/arxiv_2605_012532/requirements.txt
aiohttp>=3.9
aiosqlite>=0.20
pydantic>=2.7
ccxt>=4.3
pyyaml>=6.0
numpy>=1.26
python-dotenv>=1.0
aiofiles>=23.0
uvloop>=0.19
